+++ lib/app/services/services.dart
@@ -1,4 +1,5 @@
 export 'googlesheets_service.dart';
+export 'onboarding_service.dart';
 export 'roommate_service.dart';
 export 'service_utils.dart';
 export 'storage/hive_storage_service.dart';