+++ lib/app/services/storage/storage_key.dart
@@ -1,7 +1,7 @@
 /// Specify all local storage keys in StorageKey
 class StorageKey {
   static const String container = 'roomieTasks';
-  static const String hasOnboardingBeenShown = 'hasOnboardingBeenShown';
+  static const String onboardingCompleted = 'onboardingCompleted';
   static const String username = 'username';
   static const String spreadsheetId = 'spreadsheetId';
   static const String credentials = 'gsheetsCredentials';